+++ packages/acp-bridge/package.json
@@ -67,6 +67,10 @@
       "types": "./dist/bridgeFileSystem.d.ts",
       "import": "./dist/bridgeFileSystem.js"
     },
+    "./internal/testUtils": {
+      "types": "./dist/internal/testUtils.d.ts",
+      "import": "./dist/internal/testUtils.js"
+    },
     "./package.json": "./package.json"
   },
   "scripts": {
